+++ R/mo_property.R
@@ -974,7 +974,7 @@ mo_validate <- function(x, property, language, keep_synonyms = keep_synonyms, ..
   # try to catch an error when inputting an invalid argument
   # so the 'call.' can be set to FALSE
   tryCatch(x[1L] %in% unlist(AMR_env$MO_lookup[1, property, drop = TRUE]),
-    error = function(e) stop(e$message, call. = FALSE)
+    error = function(e) stop(conditionMessage(e), call. = FALSE)
   )
 
   dots <- list(...)